Huge input parsing / execution refacto + headers reorganization
