+++ src/server/CMakeLists.txt
@@ -4,6 +4,7 @@ include_directories(core)
 include_directories(implementation)
 include_directories(utils)
 include_directories(include)
+include_directories(include/implementation)
 
 add_executable(zappy_server
         main.c
@@ -19,19 +20,22 @@ add_executable(zappy_server
         core/welcome_single_client.c
         core/write_to_client.c
         core/spawn_resources.c
-        implementation/cmd_parser.c
-        implementation/team_parser.c
+        implementation/input_parser.c
         implementation/get_remaining_slots.c
-        implementation/gui/msz.c
-        implementation/gui/bct.c
-        implementation/gui/tna.c
-        implementation/gui/sgt.c
-        implementation/ia/forward.c
-        implementation/ia/right.c
-        implementation/ia/left.c
-        implementation/ia/take.c
-        implementation/ia/set.c
-        implementation/ia/inventory.c
+        implementation/gui/append_gui_client.c
+        implementation/gui/gui_cmd_executor.c
+        implementation/gui/cmd/msz.c
+        implementation/gui/cmd/bct.c
+        implementation/gui/cmd/tna.c
+        implementation/gui/cmd/sgt.c
+        implementation/player/append_player_client.c
+        implementation/player/player_cmd_scheduler.c
+        implementation/player/cmd/forward.c
+        implementation/player/cmd/right.c
+        implementation/player/cmd/left.c
+        implementation/player/cmd/take.c
+        implementation/player/cmd/set.c
+        implementation/player/cmd/inventory.c
         utils/concat_str.c
         utils/free_word_array.c
         utils/get_sockaddr.c
@@ -40,11 +44,13 @@ add_executable(zappy_server
         utils/revstr.c
         utils/str_to_word_array.c
         utils/word_array_len.c
-        include/server_implementation.h
-        include/server_core.h
+        include/core.h
         include/utils.h
         include/resources.h
         include/ranges.h
+        include/implementation/implementation.h
+        include/implementation/gui_cmd.h
+        include/implementation/player_cmd.h
         )
