+++ src/server/implementation/gui/cmd/sgt.c
@@ -6,7 +6,7 @@
 */
 
 #include <stdio.h>
-#include "server_implementation.h"
+#include "implementation.h"
 #include "utils.h"
 
 int do_sgt(data_t *data, char **args)
