+++ src/server/implementation/player/cmd/set.c
@@ -6,7 +6,7 @@
 */
 
 #include <stdio.h>
-#include "server_implementation.h"
+#include "implementation.h"
 #include "utils.h"
 
 int set(data_t *data, char **args)
