--- src/server/implementation/team_parser.c
@@ -1,69 +1,0 @@
-/*
-** EPITECH PROJECT, 2023
-** zappy_server
-** File description:
-** team_parser
-*/
-
-#include <string.h>
-#include <stdio.h>
-#include "server_implementation.h"
-#include "utils.h"
-
-static int append_to_team(data_t *data, char *team_name)
-{
-    int remaining_slots = get_remaining_slots(data, team_name);
-    char *str_remaining_slots = NULL;
-    char *world_dimensions = NULL;
-
-    if (remaining_slots <= 0) {
-        send_to_client(data->clients, data->curr_cli_index, "ko\n");
-        return 84;
-    }
-    init_player(&(data->clients[data->curr_cli_index]), team_name, data->map);
-    str_remaining_slots = int_to_s(remaining_slots - 1);
-    str_remaining_slots = concat_str(str_remaining_slots, "\n");
-    world_dimensions = get_world_dimensions(data);
-    send_to_client(data->clients, data->curr_cli_index, str_remaining_slots);
-    send_to_client(data->clients, data->curr_cli_index, world_dimensions);
-    free(str_remaining_slots);
-    free(world_dimensions);
-    return 0;
-}
-
-static int append_to_gui(data_t *data)
-{
-    char *x_str = NULL;
-    char *y_str = NULL;
-
-    do_msz(data, NULL);
-    do_sgt(data, NULL);
-    for (int x = 0; x < data->map->width; x++) {
-        for (int y = 0; y < data->map->height; y++) {
-            x_str = int_to_s(x);
-            y_str = int_to_s(y);
-            do_bct(data, (char *[]){x_str, y_str, NULL});
-            free(x_str);
-            free(y_str);
-        }
-    }
-    do_tna(data, NULL);
-    return 0;
-}
-
-int parse_team_name(data_t *data)
-{
-    if (!strcmp("GRAPHIC", data->clients[data->curr_cli_index]->input)) {
-        data->clients[data->curr_cli_index]->is_registered = 1;
-        return append_to_gui(data);
-    }
-    for (int i = 0; data->team_names[i]; i++) {
-        if (!strcmp(data->team_names[i],
-            data->clients[data->curr_cli_index]->input)) {
-            data->clients[data->curr_cli_index]->is_registered = 1;
-            return append_to_team(data, data->team_names[i]);
-        }
-    }
-    send_to_client(data->clients, data->curr_cli_index, "ko\n");
-    return 0;
-}
