+++ src/server/include/ranges.h
@@ -9,10 +9,27 @@
     #define ZAPPY_RANGES_H
 
     #define MAX_TEAMS 4
+
     #define MIN_MAP_SIZE 10
     #define MAX_MAP_SIZE 30
     #define TILE_SIZE 7
+
     #define MIN_FREQ 2
     #define MAX_FREQ 1000
+
+    #define MAX_PENDING_CMD 10
+
+    #define FORWARD_DELAY 7
+    #define RIGHT_DELAY 7
+    #define LEFT_DELAY 7
+    #define LOOK_DELAY 7
+    #define INVENTORY_DELAY 1
+    #define BROADCAST_DELAY 7
+    #define CONNECT_NBR_DELAY 0
+    #define FORK_DELAY 42
+    #define EJECT_DELAY 7
+    #define TAKE_DELAY 7
+    #define SET_DELAY 7
+    #define INCANTATION_DELAY 300
 
 #endif /* ZAPPY_RANGES_H */
